requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.38
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
alembic==1.14.1
gunicorn==23.0.0
slack_sdk==3.26.0 # Slack SDKを追加
line-bot-sdk==3.12.0 # LINE SDK追加
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
pytz==2024.1
